requirements.txt
flask
pillow
pytesseract
gunicorn
